# DOCS/api-documentation.yml
openapi: 3.0.3
info:
  title: ThriveSend API
  description: |
    Comprehensive API for ThriveSend platform enabling service providers to manage 
    multi-client social media campaigns, content creation, analytics, and monetization features.
  version: 1.0.0
  contact:
    name: ThriveSend API Support
    email: api-support@thrivesend.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.thrivesend.com/v1
    description: Production server
  - url: https://staging-api.thrivesend.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /content:
    get:
      tags:
        - Content
      summary: Get content list
      parameters:
        - in: query
          name: clientId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, REVIEW, APPROVED, PUBLISHED, ARCHIVED]
        - in: query
          name: type
          schema:
            type: string
            enum: [BLOG, SOCIAL, EMAIL, VIDEO, IMAGE]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Content list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Content'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Content
      summary: Create new content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentRequest'
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
  /content/{contentId}:
    get:
      tags:
        - Content
      summary: Get content details
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'

    put:
      tags:
        - Content
      summary: Update content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContentRequest'
      responses:
        '200':
          description: Content updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'

    delete:
      tags:
        - Content
      summary: Delete content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '204':
          description: Content deleted successfully

  # Calendar & Events
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ContentId:
      name: contentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Content unique identifier
...
  schemas:
...
    Content:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [BLOG, SOCIAL, EMAIL, VIDEO, IMAGE]
        status:
          type: string
          enum: [DRAFT, REVIEW, APPROVED, PUBLISHED, ARCHIVED]
        clientId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        publishedAt:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateContentRequest:
      type: object
      required:
        - title
        - body
        - type
        - clientId
      properties:
        title:
          type: string
        body:
          type: string
        type:
          type: string
          enum: [BLOG, SOCIAL, EMAIL, VIDEO, IMAGE]
        clientId:
          type: string
          format: uuid
        metadata:
          type: object
    UpdateContentRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [DRAFT, REVIEW, APPROVED, PUBLISHED, ARCHIVED]
        metadata:
          type: object
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...